feat: Checkpoint locating support for running out-of-directory
Currently, only the current directory is searched for the model checkpoint.
While manual specification of the checkpoint path is possible, this requires that you
A) remember to manually specify the path and B) know what the proper path is.

To make running outside of the installation directory easier, add a fall-back to search the project path for the checkpoint file.
This only applies only if you would otherwise get an error with the current behavior.