feat(io): Enable input from FASTA files
Add the ability for RF3 to load in from FASTA files.
The file format is inspired by Boltz's FASTA input format, but slightly more flexible.
(You should be able to input a protein FASTA as-is and have it work, albeit without MSA.)

It's written with an eye to be flexible for additional sequence file input formats, as desire dictactes.

The FASTA input is basically just syntactic sugar around the JSON input format, with a reduced feature set.